This is looking much better.
Added a better iniializer! Now you can include 2 games as a
starting position (maybe the max/min of some other game?)
This can still be improved.

Better documentation everywhere!

Changed the winner method  to return ints. To make up for it
included a function (windecode) that returns string output.

Tightened up the num function A LOT and it is MUCH FASTER and
MUCH MORE USEFUL. Still has limits and still has work to be done.